TIP-482: dissociate structure and data array conversion exceptions
