+++ src/Pim/Component/Connector/Exception/ArrayConversionException.php
@@ -2,10 +2,8 @@
 
 namespace Pim\Component\Connector\Exception;
 
-use Symfony\Component\Validator\ConstraintViolationListInterface;
-
 /**
- * Exception which may be thrown when we convert an array
+ * Exception thrown during a conversion of an array
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
@@ -13,25 +11,4 @@ use Symfony\Component\Validator\ConstraintViolationListInterface;
  */
 class ArrayConversionException extends \LogicException
 {
-    /** @var ConstraintViolationListInterface */
-    protected $violations;
-
-    public function __construct(
-        $message,
-        $code = 0,
-        \Exception $previous = null,
-        ConstraintViolationListInterface $violations = null
-    ) {
-        $this->violations = $violations;
-
-        parent::__construct($message, $code, $previous);
-    }
-
-    /**
-     * @return ConstraintViolationListInterface
-     */
-    public function getViolations()
-    {
-        return $this->violations;
-    }
 }
