+++ src/Pim/Component/Connector/Exception/StructureArrayConversionException.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Pim\Component\Connector\Exception;
+
+/**
+ * Exception thrown during a conversion of an array.
+ * Related to a structure problem, for example, an unknown property has been provided as input.
+ *
+ * @author    Julien Janvier <jjanvier@akeneo.com>
+ * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class StructureArrayConversionException extends ArrayConversionException
+{
+}
