+++ src/Pim/Component/Connector/spec/ArrayConverter/FieldsRequirementCheckerSpec.php
@@ -14,35 +14,35 @@ class FieldsRequirementCheckerSpec extends ObjectBehavior
     function it_does_not_raise_exception_when_there_is_no_required_fields()
     {
         $this
-            ->shouldNotThrow('Pim\Component\Connector\Exception\ArrayConversionException')
+            ->shouldNotThrow('Pim\Component\Connector\Exception\StructureArrayConversionException')
             ->during('checkFieldsPresence', [['foo' => 'bar'], []]);
     }
 
     function it_does_not_raise_exception_when_all_required_fields_are_filled()
     {
         $this
-            ->shouldNotThrow('Pim\Component\Connector\Exception\ArrayConversionException')
+            ->shouldNotThrow('Pim\Component\Connector\Exception\StructureArrayConversionException')
             ->during('checkFieldsPresence', [['foo' => 'bar'], ['foo']]);
     }
 
     function it_should_raise_exception_when_a_required_field_is_blank()
     {
         $this
-            ->shouldThrow('Pim\Component\Connector\Exception\ArrayConversionException')
+            ->shouldThrow('Pim\Component\Connector\Exception\DataArrayConversionException')
             ->during('checkFieldsFilling', [['foo' => ''], ['foo']]);
     }
 
     function it_should_raise_exception_when_a_required_field_is_null()
     {
         $this
-            ->shouldThrow('Pim\Component\Connector\Exception\ArrayConversionException')
+            ->shouldThrow('Pim\Component\Connector\Exception\DataArrayConversionException')
             ->during('checkFieldsFilling', [['foo' => null], ['foo']]);
     }
 
     function it_should_raise_exception_when_a_required_field_is_not_present()
     {
         $this
-            ->shouldThrow('Pim\Component\Connector\Exception\ArrayConversionException')
+            ->shouldThrow('Pim\Component\Connector\Exception\StructureArrayConversionException')
             ->during('checkFieldsPresence', [['foo' => 'bar'], ['baz']]);
     }
 }
